Checkpoint. Refactoring namespaces/names of MediatR Notification handlers.
