+++ Src/RCommon.Mediatr/Producers/PublishWithMediatREventProducer.cs
@@ -23,7 +23,6 @@ namespace RCommon.MediatR.Producers
             where T : ISerializableEvent
         {
             await _mediatorService.Publish(@event, cancellationToken);
-            Console.WriteLine("Publisher Executed w/ event: {0}", @event);
         }
     }
 }
