+++ Src/RCommon.Mediatr/Subscribers/MediatRRequest.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RCommon.Mediator.MediatR
+namespace RCommon.MediatR.Subscribers
 {
     public class MediatRRequest<TRequest, TResponse> : IRequest<TResponse>
     {
