+++ Src/RCommon.Mediatr/Subscribers/MediatRRequestHandler.cs
@@ -1,13 +1,12 @@
 ﻿using MediatR;
 using RCommon.Mediator;
-using RCommon.Mediator.MediatR;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RCommon.MediatR
+namespace RCommon.MediatR.Subscribers
 {
     public class MediatRRequestHandler<TRequest, TResponse> : IRequestHandler<MediatRRequest<TRequest, TResponse>, TResponse>
     {
